refactor: 提取格式化工具函数，添加 iOS 存储空间检查
- 将 formatBytes/formatSpeed 提取到 format_utils.dart 消除重复代码
- 添加 iOS 端 MethodChannel 实现存储空间查询
