+++ lib/utils/format_utils.dart
@@ -1,0 +1,16 @@
+String formatBytes(int bytes) {
+  if (bytes < 1024) return '$bytes B';
+  if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
+  if (bytes < 1024 * 1024 * 1024) {
+    return '${(bytes / 1024 / 1024).toStringAsFixed(1)} MB';
+  }
+  return '${(bytes / 1024 / 1024 / 1024).toStringAsFixed(1)} GB';
+}
+
+String formatSpeed(double bytesPerSec) {
+  if (bytesPerSec < 1024) return '${bytesPerSec.toStringAsFixed(0)} B/s';
+  if (bytesPerSec < 1024 * 1024) {
+    return '${(bytesPerSec / 1024).toStringAsFixed(1)} KB/s';
+  }
+  return '${(bytesPerSec / 1024 / 1024).toStringAsFixed(1)} MB/s';
+}
